+++ go/parameter_by_type_transformer.go
@@ -1,14 +1,17 @@
 package cucumberexpressions
 
 import (
-	"errors"
 	"fmt"
+	"math/big"
 	"reflect"
 	"strconv"
 )
 
 // can be imported from "math/bits". Not yet supported in go 1.8
 const uintSize = 32 << (^uint(0) >> 32 & 1) // 32 or 64
+
+const BigFloatKind = reflect.Invalid + 1000
+const BigIntKind = reflect.Invalid + 1001
 
 type ParameterByTypeTransformer interface {
 	// toValueType accepts either reflect.Type or reflect.Kind
@@ -30,7 +33,7 @@ func (s BuiltInParameterTransformer) Transform(fromValue string, toValueType int
 	return nil, createError(fromValue, toValueType)
 }
 
-func transformKind(fromValue string, toValueKind reflect.Kind) (interface{}, error) {
+func transformKind(fromValue string, toValueKind interface{String() string}) (interface{}, error) {
 	switch toValueKind {
 	case reflect.String:
 		return fromValue, nil
@@ -100,14 +103,23 @@ func transformKind(fromValue string, toValueKind reflect.Kind) (interface{}, err
 		return nil, err
 	case reflect.Float64:
 		return strconv.ParseFloat(fromValue, 64)
+	case BigFloatKind:
+		floatVal, _, err := big.ParseFloat(fromValue, 10, 1024, big.ToNearestEven)
+		return floatVal, err
+	case BigIntKind:
+		floatVal, success := new(big.Int).SetString(fromValue, 10)
+		if !success {
+			return nil, fmt.Errorf("Could not parse bigint: %s", fromValue)
+		}
+		return floatVal, nil
 	default:
 		return nil, createError(fromValue, toValueKind.String())
 	}
 }
 
 func createError(fromValue string, toValueType interface{}) error {
-	return errors.New(fmt.Sprintf("Can't transform '%s' to %s. "+
+	return fmt.Errorf("Can't transform '%s' to %s. "+
 		"BuiltInParameterTransformer only supports a limited number of types. "+
 		"Consider using a different object mapper or register a parameter type for %s",
-		fromValue, toValueType, toValueType))
+		fromValue, toValueType, toValueType)
 }